### 2021-05-05 #3251 @whedon
[PRE REVIEW]: The SAGE Rejected Article Tracker
**Submitting author:** @ad48 (<a href="http://orcid.org/0000-0002-8529-9990">Adam Day</a>)
**Repository:** <a href="https://github.com/sagepublishing/rejected_article_tracker_pkg" target ="_blank">https://github.com/sagepublishing/rejected_article_tracker_pkg</a>
**Version:** 1.5.0
