### 2021-05-06 #3251 @ad48
I noticed that our affiliation contains an apostrophe. Guessing this is the source of the problem if it's interpreted as a quote. I have now put the affiliation into double quotes to avoid ambiguity. Again, paper appears to compile correctly at https://whedon.theoj.org/ Is there anything more to do just now? Please let me know.
