+++ paper/paper.md
@@ -8,10 +8,10 @@ tags:
 authors:
   - name: Andrew Hails
     orcid: 0000-0002-1014-621X
-    affiliation: 1
+    affiliation: "1"
   - name: Adam Day
     orcid: 0000-0002-8529-9990
-    affiliation: 1
+    affiliation: "1"
 affiliations:
   - name: "SAGE Publishing, 1 Oliver's Yard, London, EC1Y 1SP"
 date: 4 November 2020
