+++ paper/paper.md
@@ -13,7 +13,7 @@ authors:
     orcid: 0000-0002-8529-9990
     affiliation: "1"
 affiliations:
-  - name: "SAGE Publishing, 1 Oliver's Yard, London, EC1Y 1SP"
+  - name: "SAGE Publishing, 1 Olivers Yard, London, EC1Y 1SP"
 date: 4 November 2020
 bibliography: paper.bib
 
