+++ paper/paper.md
@@ -8,12 +8,13 @@ tags:
 authors:
   - name: Andrew Hails
     orcid: 0000-0002-1014-621X
-    affiliation: "1"
+    affiliation: 1
   - name: Adam Day
     orcid: 0000-0002-8529-9990
-    affiliation: "1"
+    affiliation: 1
 affiliations:
-  - name: "SAGE Publishing, 1 Olivers Yard, London, EC1Y 1SP"
+  - name: "SAGE Publishing, 1 Oliver's Yard, London, EC1Y 1SP"
+  - index: 1
 date: 4 November 2020
 bibliography: paper.bib
 
