+++ rejected_article_tracker/src/RejectedArticlesMatch.py
@@ -12,6 +12,7 @@ from .EmptyResult import EmptyResult
 
 from .ML.config import Config as mlconfig
 
+from .LoadModel import LoadModel
 
 class RejectedArticlesMatch:
 
@@ -28,21 +29,8 @@ class RejectedArticlesMatch:
         self.email = email
         self.config = config
         self.results = results
-        self.clf = self.load_model()
+        self.clf = LoadModel().clf
 
-    def load_model(self):
-        # if we have created a new model using the training functions
-        # then this will pick that model instead of the standard one.
-        new_model_path = mlconfig.new_logreg_model_loc
-        old_model_path = mlconfig.old_logreg_model_loc
-        if os.path.exists(new_model_path):
-            with open(new_model_path, 'rb') as f:
-                clf = pickle.load(f)
-        else:
-            with open(old_model_path, 'rb') as f:
-                clf = pickle.load(f)
-        return clf
-        
 
     def match(self) -> list:
         """
