+++ rejected_article_tracker/src/ML/config.py
@@ -24,7 +24,7 @@ class Config:
 
     # model locations
     # file_dir = os.path.dirname(__file__)
-    ml_model_dir = os.path.join(user_home_dir,'rejected_article_tracker/models')
+    ml_model_dir = os.path.join(user_home_dir,'rejected_article_tracker','models')
     old_logreg_model_loc = os.path.join(ml_model_dir, 'lr_model')
     new_logreg_model_loc = os.path.join(ml_model_dir, 'lr_model_new')
      
