### 2021-05-25 90a5042
improve handling of lr_model file
