### 2021-06-08 #3251 @danielskatz
👋 @jasonpriem, @hpiwowar, @mfenner - Would one of you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
