### 2021-06-09 #3251 @ad48
@danielskatz I looked at the spreadsheet and found these names who seem like they would have the right skills and interests.

-- Bibliometrics + Python
gagolews
(there are a few others with Bibliometrics / Scientometrics in their skills, but not Python)

-- ScholComms + Python
dhimmel

-- NLP + Python
bagheria
indika.kahanda at montana
(the NLP in our application is very basic, so on the one hand a reviewer with NLP skills might help improve it, on the other they might not find much of interest, so I'm not sure if this is the right way to go)
