### 2021-06-09 #3251 @danielskatz
👋 @gagolews, @dhimmel  - Would one of you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
