### 2021-06-09 #3251 @jodischneider
Sorry, I'm overbooked - but feel free to ask me for another scholcomm
project in the future, or if you lose a quorum of reviewers on this one!

On Tue, Jun 8, 2021 at 1:31 PM Daniel S. Katz ***@***.***>
wrote:

> 👋 @jodischneider <https://github.com/jodischneider> - Would you (or
> someone in your group) be willing to review this submission to JOSS? We
> carry out our checklist-driven reviews here in GitHub issues and follow
> these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
