### 2021-06-15 sagepublishing/rejected_article_tracker_pkg#4 @dhimmel
CI build is failing: missing tqdm
https://github.com/sagepublishing/rejected_article_tracker_pkg/runs/2785999220#step:5:22

```python-traceback
ERROR: rejected_article_tracker (unittest.loader._FailedTest)
ImportError: Failed to import test module: rejected_article_tracker
Traceback (most recent call last):
  File "/opt/hostedtoolcache/Python/3.9.5/x64/lib/python3.9/unittest/loader.py", line 470, in _find_test_path
    package = self._get_module_from_name(name)
  File "/opt/hostedtoolcache/Python/3.9.5/x64/lib/python3.9/unittest/loader.py", line 377, in _get_module_from_name
    __import__(name)
  File "/home/runner/work/rejected_article_tracker_pkg/rejected_article_tracker_pkg/rejected_article_tracker/__init__.py", line 11, in <module>
    from .src.ML.ArXivOAIPMH import ArXivOAIPMH
  File "/home/runner/work/rejected_article_tracker_pkg/rejected_article_tracker_pkg/rejected_article_tracker/src/ML/ArXivOAIPMH.py", line 4, in <module>
    from tqdm import tqdm
ModuleNotFoundError: No module named 'tqdm'
```
