### 2021-06-15 #3348 @dhimmel
# Daniel Himmelstein Review

Draft: will keep updating this comment.

Here's a list of issues related to my review:

- https://github.com/sagepublishing/rejected_article_tracker_pkg/issues/4
