### 2021-06-21 sagepublishing/rejected_article_tracker_pkg#4 @ad48
Thanks for pointing this out. tqdm is present in setup.py now. I think that resolves the issue.
