### 2021-06-21 sagepublishing/rejected_article_tracker_pkg#4 @danielskatz
in https://github.com/openjournals/joss-reviews/issues/3348
