### 2021-06-21 sagepublishing/rejected_article_tracker_pkg#4 @dhimmel
> tqdm is present in setup.py now

Ah I do see https://github.com/sagepublishing/rejected_article_tracker_pkg/commit/bc2efec008c2c6fbf9d5d08a3f7926701ed74d26 that added tqdm, but this seems to have predated this issue.

I'm not sure why fc4f72de7b12c4acc6af597d61464ef7ed6e943b has that CI failure then. Anyways looking through the commit history, it looks like CI has been failing for a long time. CI on main should be passing.
