### 2021-06-21 sagepublishing/rejected_article_tracker_pkg#4 @dhimmel
Ah I do I know why. The dependencies are defined in two places: `requirements.txt` and `setup.py`. `requirements.txt` is used by CI and is missing `tqdm`:

https://github.com/sagepublishing/rejected_article_tracker_pkg/blob/fc4f72de7b12c4acc6af597d61464ef7ed6e943b/requirements.txt#L1-L10

Would probably be best to single-source your dependencies, although Python is not great at that. Noting https://packaging.python.org/discussions/install-requires-vs-requirements/. Is there any reason you need `requirements.txt`? Could probably just get by with `install_requires`?
