### 2021-06-21 sagepublishing/rejected_article_tracker_pkg#4 @dhimmel
While the general convention is that maintainers close issues, for the issues related to my openjournals/joss-reviews#3348 review, should I close my own issues? This will make it so we can use issue open/close status as an indicator of my satisfaction regarding a specific point?
