### 2021-06-21 sagepublishing/rejected_article_tracker_pkg#4 @danielskatz
either way is fine - you can always reopen the issue or open a new one if @ad48 closes something you disagree with
