### 2021-06-22 ce3fbce
update requirements to match setup see: issues/4
