### 2021-06-22 sagepublishing/rejected_article_tracker_pkg#4 @ad48
Apologies - I may have closed the issue prematurely. I have now updated the requirements.txt file.
