### 2021-06-22 c30b755
update re: issues 6,7,8
