+++ paper/paper.bib
@@ -159,6 +159,12 @@ title = {The STM Report},
 year = {2018},
 url = {https://www.stm-assoc.org/2018_10_04_STM_Report_2018.pdf}
 }
+@online{ScholarOne,
+author = {Clarivate},
+title = {Workflow-management systems for scholarly journals - ScholarOne - Web of Science Group},
+year = {2021},
+url = {https://clarivate.com/webofsciencegroup/solutions/scholarone/}
+}
 @misc{SAGERAT,
 title = {SAGE Rejected Article Tracker},
 author = {Hails, Andrew; Day, Adam},
