### 2021-06-28 9a99b2d
handle missing MY_EMAIL environment variable
