### 2021-06-28 sagepublishing/rejected_article_tracker_pkg#4 @dhimmel
> I have now updated the requirements.txt file.

Got it! I see https://github.com/sagepublishing/rejected_article_tracker_pkg/commit/ce3fbce99bdc4e7b75b9206ddb15566aabddc66a. Thanks for mentioning the issue in the commit `see: issues/4`. If you format this like `#4`, then GitHub will automatically show the commit on this issue.

CI is [now failing](https://github.com/sagepublishing/rejected_article_tracker_pkg/runs/2913467719) due to a different issue:

```python-traceback
  File "/home/runner/work/rejected_article_tracker_pkg/rejected_article_tracker_pkg/rejected_article_tracker/src/CrossRef.py", line 46, in search
    'mailto': os.environ['MY_EMAIL']
  File "/opt/hostedtoolcache/Python/3.9.5/x64/lib/python3.9/os.py", line 679, in __getitem__
    raise KeyError(key) from None
KeyError: 'MY_EMAIL'
```

Perhaps you need to set an email for testing?
