### 2021-06-29 #3348 @ad48
@danielskatz my coauthor and I have been discussing it and we've made a few changes. We resolved all but 1 of the issues. (Should I close those ones now?)
