### 2021-06-29 sagepublishing/rejected_article_tracker_pkg#4 @ad48
Ah thanks! I had missed that. I've corrected this and github actions seem to now be working properly.

Thanks for the tip about issue numbering. I will try to keep that in mind in future.
