### 2021-06-29 sagepublishing/rejected_article_tracker_pkg#4 @dhimmel
> I've corrected this and github actions seem to now be working properly.

Noting https://github.com/sagepublishing/rejected_article_tracker_pkg/commit/9a99b2dd33cbf117b05b5a800ca2de8209f31259, which [passed CI](https://github.com/sagepublishing/rejected_article_tracker_pkg/runs/2940801327)!
