### 2021-06-29 #3348 @dhimmel
> we've made a few changes

Ah I hadn't seen https://github.com/sagepublishing/rejected_article_tracker_pkg/commit/c30b755619bf0e9517f848827911797de5e37275 with changes tot paper, will take a look at that and update the issues.
