### 2021-07-03 #3348 @dhimmel
Went through my open issues and closed the resolved ones. Still one remaining.
