### 2021-07-05 a55182c
update paper with 'how it works' re issue#7
