+++ paper/paper.md
@@ -30,7 +30,7 @@ Given metadata for a rejected article, the rejected article tracker will:
 
 * select the most likely correct result from that list using simple machine learning.
 
-The target audience for the tracker is researchers studying rejected articles, or performing DOI-resolution for other reasons, such as assigning DOIs to preprints in order to connect the preprints to their published versions. 
+The target audience for the tracker is researchers studying rejected articles. The task performed by the tracker is 'DOI-resolution'. I.e. finding the correct Digital Object Identifier given incomplete data about a paper. So, while the intended use of the tracker is to track rejected articles, it might be used by researchers performing DOI-resolution for other reasons, such as assigning DOIs to preprints in order to connect the preprints to their published versions. This is a particularly topical application at the current time due to the rapid growth of preprint servers in recent years. [REFERENCE]
 
 The tracker is available as [a Python package](https://github.com/ad48/rejected_article_tracker_pkg) with [a temporary live demonstration](https://rejectedarticlestorage.z6.web.core.windows.net/) scheduled to run until mid-2021.
 
@@ -57,6 +57,41 @@ Common forms of author-misconduct can be identified and studied.
 Finally, the rejected article tracker can also be used to link preprints with their published versions. Due to the rapid recent growth in preprint servers [@Fraser2020], there is a growing need to improve the data-quality surrounding preprints. 
 
 The rejected article tracker is set up, by default, to accept data in the format exported by [ScholarOne](https://clarivate.com/webofsciencegroup/solutions/scholarone/), presently the most popular system for managing peer-review. However, the input data required is minimal, so data from any peer-review management system should be easily adapted for the tracker. Instructions are given in the Readme.md file of [the github repository](https://github.com/ad48/rejected_article_tracker_pkg).
+
+## How the matching algorithm works
+
+The CrossRef API is often used to perform DOI resolution. A typical use case would be for the purpose of adding DOIs and other data to incomplete references in reference-lists. Under this typical use-case, there is often other data available, such as journal name and publication date as well as issue, volume or page numbers. However, if we wish to track rejected articles, it is likely that we only have the title and author-names for an article and there is a lower chance that it exists in CrossRef's data (since not all rejected articles are published). So, searching the API for just these 2 things often results in incorrect results being retrieved.
+
+We start with a dataset of ArXiv preprint metadata retrieved from the [ArXiv OAI-PMH API](https://arxiv.org/help/oa/index). This dataset resembles journal submission data in that it includes the titles and author names of preprints. In many cases, this data also includes the DOI of the same article when it was published. This means that we know the correct result of a DOI-resolution process for this article. We find that the title and author lists are not always identical. Titles often undergo minor (and occasionally major) changes and author lists can also change in a number of ways (full names might be used instead of initials, or perhaps new authors are added to an author list at some point in the process).
+
+We search the CrossRef API for each preprint's DOI as well as the best incorrect search result. This means that we can fill out 2 rows of a table of data for each preprint.
+
+| preprint title | preprint authors | published title | published authors | correct/incorrect |
+|-|-|-|-|-|
+| title1| author_list1 | title2 | author_list2 | correct |
+| title1| author_list1 | title3 | author_list3 | incorrect |
+
+We then 
+- Calculate the Levenshtein distance between the titles in each row. This is normalised to a number between 0 and 100 using the `fuzzywuzzy` package's `fuzz.ratio` method.
+- Normalise all author names to a single string of `first_initial+last_name` in lower case. Then calculate 2 boolean values: one showing if there is a 100% match in author lists and one showing if there is at least 1 author name matching in the 2 lists. 
+
[…]
