+++ paper/paper.bib
@@ -165,6 +165,13 @@ title = {Workflow-management systems for scholarly journals - ScholarOne - Web o
 year = {2021},
 url = {https://clarivate.com/webofsciencegroup/solutions/scholarone/}
 }
+
+@online{Scikit-Learn,
+author = {Scikit-Learn},
+title = {sklearn.linear_model.LogisticRegression - scikit-learn 0.24.2 documentation},
+year = {2020},
+url = {https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LogisticRegression.html}
+}
 @misc{SAGERAT,
 title = {SAGE Rejected Article Tracker},
 author = {Hails, Andrew; Day, Adam},
