### 2021-07-06 94ab091
update paper re: issue#7
