+++ paper/paper.bib
@@ -139,6 +139,44 @@ journal = {PLoS Biol 19(4): e3000959},
 url = {https://doi.org/10.1371/journal.pbio.3000959},
 doi = {10.1371/journal.pbio.3000959}
 }
+@article{Cabanac2021,
+author = {Cabanac, Guillaume; Oikonomidi, Theodora; Boutron, Isabelle},
+title = {Day-to-day discovery of preprint–publication links},
+journal = {Scientometrics},
+issn = {1588-2861},
+volume = {126},
+number = {6},
+pages = {5285-5304},
+year  = {2020},
+publisher = {Springer Nature},
+doi = {10.1007/s11192-021-03900-7},
+URL = { 
+        https://doi.org/10.1007/s11192-021-03900-7
+},
+}
+@article{Hoy2020,
+author = {Matthew B. Hoy},
+title = {Rise of the Rxivs: How Preprint Servers are Changing the Publishing Process},
+journal = {Medical Reference Services Quarterly},
+volume = {39},
+number = {1},
+pages = {84-89},
+year  = {2020},
+publisher = {Routledge},
+doi = {10.1080/02763869.2020.1704597},
+    note ={PMID: 32069196},
+
+URL = { 
+        https://doi.org/10.1080/02763869.2020.1704597
+    
+},
+eprint = { 
+        https://doi.org/10.1080/02763869.2020.1704597
+    
+}
+
+}
+
 @online{HighWire,
 author = {HighWire},
 title = {HighWire Analytics - Highwire Press},
