+++ paper/paper.md
@@ -60,7 +60,7 @@ The rejected article tracker is set up, by default, to accept data in the format
 
 ## How the matching algorithm works
 
-The CrossRef API is often used to perform DOI resolution. A typical use case would be for the purpose of adding DOIs and other data to incomplete references in reference-lists. Under this typical use-case, there is often other data available, such as journal name and publication date as well as issue, volume or page numbers. However, if we wish to track rejected articles, it is likely that we only have the title and author-names for an article and there is a lower chance that it exists in CrossRef's data (since not all rejected articles are published). So, searching the API for just these 2 things often results in incorrect results being retrieved.
+The CrossRef API is often used to perform DOI resolution. A typical use case would be for the purpose of adding DOIs and other metadata to incomplete references in the reference-list of a research paper. Under this typical use-case, there is often other data available, such as journal name and publication date as well as issue, volume or page numbers. However, if we wish to track rejected articles, it is likely that we only have the title and author-names for an article and there is a lower chance that it exists in CrossRef's data (since not all rejected articles are published). So, searching the API for just these 2 things often results in incorrect results being retrieved.
 
 We start with a dataset of ArXiv preprint metadata retrieved from the [ArXiv OAI-PMH API](https://arxiv.org/help/oa/index). This dataset resembles journal submission data in that it includes the titles and author names of preprints. In many cases, this data also includes the DOI of the same article when it was published. This means that we know the correct result of a DOI-resolution process for this article. We find that the title and author lists are not always identical. Titles often undergo minor (and occasionally major) changes and author lists can also change in a number of ways (full names might be used instead of initials, or perhaps new authors are added to an author list at some point in the process).
 
