### 2021-07-07 sagepublishing/rejected_article_tracker_pkg#9 @dhimmel
FileNotFoundError: rejected_article_tracker/src/ML/small_models/lr_model
I did `pip install rejected_article_tracker` in a fresh environment. Hit the following error on import:

```python-traceback
>>> from rejected_article_tracker import RejectedArticlesMatch
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/dhimmel/miniconda3/envs/rejected-article-tracker/lib/python3.9/site-packages/rejected_article_tracker/__init__.py", line 22, in <module>
    shutil.copy(lr_model_path, dest_model_path)
  File "/home/dhimmel/miniconda3/envs/rejected-article-tracker/lib/python3.9/shutil.py", line 418, in copy
    copyfile(src, dst, follow_symlinks=follow_symlinks)
  File "/home/dhimmel/miniconda3/envs/rejected-article-tracker/lib/python3.9/shutil.py", line 264, in copyfile
    with open(src, 'rb') as fsrc, open(dst, 'wb') as fdst:
FileNotFoundError: [Errno 2] No such file or directory: '/home/dhimmel/miniconda3/envs/rejected-article-tracker/lib/python3.9/site-packages/rejected_article_tracker/src/ML/small_models/lr_model'
```

I see the following lines intended to include this file/directory in the distribution:

https://github.com/sagepublishing/rejected_article_tracker_pkg/blob/94ab0918d965437b4919ab60206e7762c40b77e8/setup.py#L17-L18

_testing as part of https://github.com/openjournals/joss-reviews/issues/3348_
