### 2021-07-07 sagepublishing/rejected_article_tracker_pkg#9 @dhimmel
Noting that [`src/ML/small_models/lr_model`](https://github.com/sagepublishing/rejected_article_tracker_pkg/blob/94ab0918d965437b4919ab60206e7762c40b77e8/rejected_article_tracker/src/ML/small_models/lr_model) is a file.

Documentation at https://setuptools.readthedocs.io/en/latest/userguide/datafiles.html.
