### 2021-07-07 sagepublishing/rejected_article_tracker_pkg#9 @ad48
Apologies! This issue came up a few weeks ago and I thought I had fixed it. Will sort it out.
