### 2021-07-07 7eac356
minor: fix bullet formatting
