### 2021-07-07 82771da
minor tweaks to paper
