+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include rejected_article_tracker/src/lr_model
+include rejected_article_tracker/src/ML/small_models/lr_model
