### 2021-07-09 sagepublishing/rejected_article_tracker_pkg#10 @mfenner
Suggested reference
Highly recommend to include (and discuss in `State of the Field`) this reference:

Dominika Tkaczyk. „Reference Matching: For Real This Time“. Website. Crossref. Zugegriffen 9. Juli 2021. https://www.crossref.org/blog/reference-matching-for-real-this-time/.

openjournals/joss-reviews#3348
