### 2021-07-09 sagepublishing/rejected_article_tracker_pkg#11 @mfenner
State of the field
It would strengthen the paper (the quality of the writing) if the text makes clearer that rejected article tracking is only one important application of the software, where the core functionality is finding DOI records in Crossref using metadata such as title and authors.

openjournals/joss-reviews#3348
