### 2021-07-09 sagepublishing/rejected_article_tracker_pkg#12 @mfenner
Example usage
Example usage is described in the readme, but a dedicated example file would make this easier.

openjournals/joss-reviews#3348
