### 2021-07-09 sagepublishing/rejected_article_tracker_pkg#13 @mfenner
Discuss external dependencies
A stronger discussion of ScholarOne as external dependency would be helpful. It is a very popular journal submission system, but more guidance for users of other submission systems would be helpful.

This could for example be done by giving one or more examples (in particular name formatting) in the section `Article format`.

openjournals/joss-reviews#3348
