### 2021-07-12 sagepublishing/rejected_article_tracker_pkg#9 @ad48
I updated `Manifest.in` and this seems to have corrected the issue. I've reinstalled and run from scratch. All appears to be working. (Sorry, I forgot to include the issue number in the commit) https://github.com/sagepublishing/rejected_article_tracker_pkg/blob/main/MANIFEST.in
