+++ paper/paper.bib
@@ -197,6 +197,12 @@ title = {The STM Report},
 year = {2018},
 url = {https://www.stm-assoc.org/2018_10_04_STM_Report_2018.pdf}
 }
+@online{Tkaczyk2018,
+author = {Tkaczyk, Dominika},
+title = {Reference matching: for real this time},
+year = {2018},
+url = {https://www.crossref.org/blog/reference-matching-for-real-this-time}
+}
 @online{ScholarOne,
 author = {Clarivate},
 title = {Workflow-management systems for scholarly journals - ScholarOne - Web of Science Group},
