### 2021-07-12 sagepublishing/rejected_article_tracker_pkg#11 @ad48
Thank you for the feedback. We added this paragraph to the paper after initial feedback from the other referee. Does this address your point? Happy to make further changes if not:

> The target audience for the tracker is researchers studying rejected articles. The task performed by the tracker is 'DOI-resolution'. I.e. finding the correct Digital Object Identifier given incomplete data about a paper. So, while the intended use of the tracker is to track rejected articles, it might be used by researchers performing DOI-resolution for other reasons, such as assigning DOIs to preprints in order to connect the preprints to their published versions, e.g. [@Cabanac2021]. This is a particularly topical application at the current time due to the rapid growth of preprint servers in recent years [@Hoy2020].
