### 2021-07-12 796af49
Add detail on name formatting re: issue#13
