### 2021-07-12 sagepublishing/rejected_article_tracker_pkg#13 @ad48
Thank you for your comment. I've updated the `readme.md` to this effect.
