### 2021-07-12 sagepublishing/rejected_article_tracker_pkg#11 @mfenner
The comment addresses my point, thank you.  I suggest one small change: DOI resolution is in the PID community understood as looking up the URL associated with the DOI, not finding a DOI for a given set of metadata. Could you use a different term?
