### 2021-07-13 sagepublishing/rejected_article_tracker_pkg#11 @ad48
Yes, I was thinking about this already. I can see the potential for confusion.

- I'm not sure that the 2 definitions are mutually exclusive. E.g. if we have a URL, we are checking to see if that information refers to a real DOI. The same is true in this process where we have incomplete metadata and we are checking to see if that information refers to a real DOI. This seems to fit with definitions for entity-resolution found with Google, e.g. : https://www.sciencedirect.com/topics/computer-science/entity-resolution but that doesn't mean there's no potential for confusion, which I would rather avoid.
- If we change the terminology, I'm not sure what would be the best thing to use. In the reference you recommended, the process is called 'reference-matching', but this isn't appropriate for rejected article tracking, because we don't have 'references' as input.

Did you have an alternative phrasing in mind? Happy to change it, just not sure what to!
