### 2021-07-13 sagepublishing/rejected_article_tracker_pkg#12 @ad48
Yes, I think we can do this. Just to check I've understood correctly, do you mean for us to create a file (e.g. JSON or xlsx) with the expected input so that users can run a test with it?
