### 2021-07-13 sagepublishing/rejected_article_tracker_pkg#11 @mfenner
DOI resolution is engrained in the DOI community, that I would avoid the term used in a different context.

Maybe "DOI lookup for unstructured metadata"?
