### 2021-07-13 sagepublishing/rejected_article_tracker_pkg#12 @mfenner
Yes, exactly.
