### 2021-07-16 513d8e1
remove doi-resolution confusion re: issue#11
