### 2021-07-16 sagepublishing/rejected_article_tracker_pkg#11 @ad48
Thanks again for the feedback. I've changed this to 'record-linkage' in the paper.
