### 2021-07-16 sagepublishing/rejected_article_tracker_pkg#9 @dhimmel
> I updated Manifest.in and this seems to have corrected the issue

Okay updated in https://github.com/sagepublishing/rejected_article_tracker_pkg/commit/591c0db717541e480bb5f37a07130badf1f34dd9. Let me know when a new version is released on [PyPI](https://pypi.org/project/rejected-article-tracker/) and I will test things again.
