### 2021-07-20 #3348 @ad48
@danielskatz  We have 2 things left on the todo list: one is to create a sample file to make testing the app easy. The other is to post the training data to an open data repository. Both are quite do-able, but time has been a little limited over this last week. Hopefully will have it done this week. The feedback from both referees has been highly constructive and greatly appreciated.
