### 2021-07-23 sagepublishing/rejected_article_tracker_pkg#12 @ad48
quick update: I'm working on this, but I thought it might be better to make it executable from the command line, so it means adding a bit more code.
