### 2021-07-26 sagepublishing/rejected_article_tracker_pkg#12 @ad48
I've added the file now as well as a simple `run_sample.py` file which processes it and produces an output file.
