+++ paper/paper.bib
@@ -224,3 +224,12 @@ publisher={GitHub},
 journal = {GitHub repository},
 url = {https://github.com/ad48/rejected_article_tracker_pkg}
 }
+@misc{SAGERATData,
+author = {Adam Day},
+year = {2021},
+title = {SAGE Rejected Article Tracker Training Data},
+version = {20210722a},
+publisher = {Zenodo},
+url =  {http://doi.org/10.5281/zenodo.5122848},
+doi = {10.5281/zenodo.5122848}
+}
