### 2021-07-26 sagepublishing/rejected_article_tracker_pkg#12 @danielskatz
So can we close this issue now?  @mfenner? (who might be on vacation this week)
