### 2021-07-26 #3348 @dhimmel
I'm waiting on a new PyPI release as noted at https://github.com/sagepublishing/rejected_article_tracker_pkg/issues/9#issuecomment-881663829
