### 2021-07-26 sagepublishing/rejected_article_tracker_pkg#12 @mfenner
Yes.
