### 2021-07-28 sagepublishing/rejected_article_tracker_pkg#9 @ad48
@dhimmel I think that's the latest version up there now. I've reinstalled from pypi and run the run_sample.py file which is working as expected.
