### 2021-07-28 #3348 @ad48
I've pushed the latest version to pypi now. Running `pip install rejected_article_tracker` will install version 1.5.6, then `python run_sample.py` produces an `output.xlsx` file which demonstrates the process working.

For context:
- The output file actually shows that the tracker fails to find one of the example articles. I thought that was worth leaving in so that users know what a null result looks like.
- We don't typically share data on real rejected articles because there are some data-protection issues in doing so.  The examples are all fake - they are real papers that are very highly cited and therefore unlikely to have been rejected from anywhere.
