### 2021-07-31 sagepublishing/rejected_article_tracker_pkg#14 @dhimmel
Is it because the example article doesn't use the article format later in the REAMDE? It uses `Manuscript ID` instead of `manuscript_id`?
