### 2021-08-02 #3348 @dhimmel
Still waiting to check off the following bullet:

>  Functionality: Have the functional claims of the software been confirmed?

Opened https://github.com/sagepublishing/rejected_article_tracker_pkg/issues/14 recently when the README code failed.
