### 2021-08-02 30bacd2
Fix example articles re: issue #14
