### 2021-08-02 sagepublishing/rejected_article_tracker_pkg#14 @ad48
Ah, well spotted! I've corrected this now and run the examples from the `readme.md` file. All working.
