### 2021-08-02 #3348 @dhimmel
> What about the two items in your checklist that are not checked?

Looked over the guidance for "Substantial scholarly effort" and "Community guidelines" and decided that I can check these off.
