+++ paper/paper.bib
@@ -1,7 +1,5 @@
 @article{Incorvia2015,
 author = {Incorvia, Currivan},
-file = {:C$\backslash$:/Users/aday/Documents/Papers/US10255359.pdf:pdf},
-mendeley-groups = {Rejected article tracking},
 number = {12},
 title = {{( 12 ) United States Patent}},
 volume = {2},
@@ -19,10 +17,11 @@ mendeley-groups = {Rejected article tracking},
 number = {3},
 pages = {450--454},
 pmid = {20099256},
-title = {{Fate of manuscripts declined by the British journal of surgery}},
+title = {{Fate of manuscripts declined by the British Journal of Surgery}},
 volume = {97},
 year = {2010}
 }
+
 @article{Citerio2018,
 author = {Citerio, Giuseppe and Deutsch, Ellen and Sala, Elisa and Lavillonni{\`{e}}re, Martin and Perner, Anders and Jaber, Samir and Timsit, Jean Francois and Azoulay, Elie},
 doi = {10.1007/s00134-018-5407-2},
@@ -37,6 +36,7 @@ title = {{Fate of manuscripts rejected by Intensive Care Medicine from 2013 to 2
 volume = {44},
 year = {2018}
 }
+
 @article{Donovan2007,
 abstract = {If your research paper is rejected for publication by a peer-reviewed journal, what do you do next? Some editors would suggest that it should not be sent to another journal, yet there are contributors who would just put a failed' paper straight back in the mail to another journal and trust to luck. I argue that rejection should be a time for reassessment, when the input of the editor(s) and reviewers(s), all experts in their fleld, should be considered and acted upon as necessary. The revitalized and improved contribution can then be sent to a new journal with confidence.},
 author = {Donovan, Stephen K.},
@@ -51,6 +51,7 @@ title = {{The importance of resubmitting rejected paper}},
 volume = {38},
 year = {2007}
 }
+
 @article{Docherty2017,
 author = {Docherty, A. B. and Klein, A. A.},
 doi = {10.1111/anae.13829},
@@ -66,6 +67,7 @@ title = {{The fate of manuscripts rejected from Anaesthesia}},
 volume = {72},
 year = {2017}
 }
+
 @article{Chung2020,
 author = {Chung, Sungjin and Lee, Jeonghwan and Yoo, Tae Hyun and Kim, Gheun Ho},
 doi = {10.23876/j.krcp.20.392},
@@ -79,6 +81,7 @@ title = {{The fate of manuscripts rejected from kidney research and clinical pra
 volume = {39},
 year = {2020}
 }
+
 @article{Hesselmann2017,
 author = {Felicitas Hesselmann and Verena Graf and Marion Schmidt and Martin Reinhart},
 title ={The visibility of scientific misconduct: A review of the literature on retracted journal articles},
@@ -117,6 +120,7 @@ eprint = {
     url = {https://doi.org/10.1093/ije/dyz187},
     eprint = {https://academic.oup.com/ije/article-pdf/49/1/281/32995499/dyz187.pdf},
 }
+
 @article{Bozzo2017,
 	title = {Retractions in cancer research: a systematic survey},
 	volume = {2},
@@ -131,14 +135,20 @@ eprint = {
 	year = {2017},
 	pages = {5}
 }
+
 @article{Fraser2020,
-author = {Fraser N, Brierley L, Dey G, Polka JK, Pálfy M, Nanni F, et al.},
-year = {2021},
-title = {The evolving role of preprints in the dissemination of COVID-19 research and their impact on the science communication landscape.},
[…]
