### 2021-08-02 #3348 @danielskatz
@ad48 - I've suggested some changes in https://github.com/sagepublishing/rejected_article_tracker_pkg/pull/15.  In addition, the table that follows line 73 overruns the margin - perhaps you can shorten some of the headers or make them two lines to avoid this. Also, the reference to Incorvia2015 seems incomplete.
