+++ paper/paper.bib
@@ -1,16 +1,14 @@
-@article{Incorvia2015,
-author = {Incorvia, Currivan},
-number = {12},
-title = {{( 12 ) United States Patent}},
-volume = {2},
-year = {2015}
+@article{Dimensions2021,
+author = {Selmani, Nora},
+title = {Using Dimensions to Track and Report on Rejected Submissions},
+url = {https://www.dimensions.ai/blog/using-dimensions-to-track-and-report-on-rejected-submissions/},
+year = {2021}
 }
 
 @article{Wijnhoven2010,
 abstract = {Background: The fate of papers submitted and subsequently rejected by the British Journal of Surgery (BJS) is currently unknown. The present study was designed to investigate whether, when and where these papers are published following rejection. Methods: All rejected manuscripts in the year 2006 were identified from the Manuscript Central electronic database. Between December 2008 and February 2009, a PubMed search was conducted spanning the period 2006-2009 using the corresponding author's last name and initials to identify whether and when manuscripts had been published elsewhere. Results: From the 926 manuscripts rejected by BJS, 609 (65-8 per cent) were published in 198 different journals with a mean(s.d.) time lapse of 13-8(6-5) months. Some 165 manuscripts (27-1 per cent) were published in general surgical journals, 250 (41-1 per cent) in subspecialty surgical journals and 194 (31-9 per cent) in non-surgical journals. The mean(s.d.) impact factor of the journals was 2-0(1-1). Only 14 manuscripts (2-3 per cent) were published in journals with a higher impact factor than that of BJS. Conclusion: Rejection of a manuscript by BJS does not preclude publication, but rejected manuscripts are published more often in surgical subspecialty journals and journals with a lower impact factor, although the occasional exception exists. Copyright {\textcopyright} 2010 British Journal of Surgery Society Ltd.},
 author = {Wijnhoven, B. P.L. and Dejong, C. H.C.},
 doi = {10.1002/bjs.6880},
-file = {:C$\backslash$:/Users/aday/Documents/Papers/bjs.6880.pdf:pdf},
 issn = {00071323},
 journal = {British Journal of Surgery},
 mendeley-groups = {Rejected article tracking},
@@ -25,7 +23,6 @@ year = {2010}
 @article{Citerio2018,
 author = {Citerio, Giuseppe and Deutsch, Ellen and Sala, Elisa and Lavillonni{\`{e}}re, Martin and Perner, Anders and Jaber, Samir and Timsit, Jean Francois and Azoulay, Elie},
 doi = {10.1007/s00134-018-5407-2},
-file = {:C$\backslash$:/Users/aday/Documents/Papers/2018 Citerio et al., fate of manuscripts.pdf:pdf},
 issn = {14321238},
 journal = {Intensive Care Medicine},
 mendeley-groups = {Rejected article tracking},
@@ -41,7 +38,6 @@ year = {2018}
 abstract = {If your research paper is rejected for publication by a peer-reviewed journal, what do you do next? Some editors would suggest that it should not be sent to another journal, yet there are contributors who would just put a failed' paper straight back in the mail to another journal and trust to luck. I argue that rejection should be a time for reassessment, when the input of the editor(s) and reviewers(s), all experts in their fleld, should be considered and acted upon as necessary. The revitalized and improved contribution can then be sent to a new journal with confidence.},
 author = {Donovan, Stephen K.},
[…]
