### 2021-08-03 #3348 @ad48
@danielskatz I have tried adding linebreaks to the column headings, but it seems whedon couldn't render those. So I have changed the headings to say 'ArXiv' and 'VOR' with some more explanatory text about what VOR means in the preceding paragraph. I think this gets around the problem with overrunning the margins.

I also removed the 'Incorvia2015' reference. That was supposed to reference the patent for the HighWire tool in the other reference. Thinking about it, the patent isn't particularly relevant given the context of the reference which is only pointing out that commercial alternatives exist. In the recent weeks (since we submitted this paper), Dimensions have started promoting their rejected article tracking service. So we have added a link to that since it's another relevant example.
