### 2021-08-03 #3348 @ad48
- @danielskatz @mfenner @dhimmel I have made a minor amendment to the acknowledgements section to include all of you. (I assume that's OK - I couldn't find guidelines on this.) I wanted to pass on our thanks to you for all of your help with this.
- Tagged release 1.5.8
- Archived in Zenodo
- Metadata appears to be correct (including ORCIDs)
- Zenodo DOI: 10.5281/zenodo.5155973
