### 2021-08-03 #3348 @danielskatz
If you can change the ack to list me as editor rather than reviewer, please do - you don't need to create a new release or update the archive, as we don't even require the paper to be in the archive, but if you want to, you can.  Let me know either way, and I'll move on the next step.
